WLTS-29 [WIP] conditional security config
