+++ src/main/java/uk/gov/hmcts/reform/translate/config/SecurityConfiguration.java
@@ -1,95 +1,47 @@
 package uk.gov.hmcts.reform.translate.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.oauth2.core.DelegatingOAuth2TokenValidator;
 import org.springframework.security.oauth2.core.OAuth2TokenValidator;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.jwt.JwtDecoder;
 import org.springframework.security.oauth2.jwt.JwtDecoders;
-import org.springframework.security.oauth2.jwt.JwtIssuerValidator;
 import org.springframework.security.oauth2.jwt.JwtTimestampValidator;
 import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter;
-import org.springframework.security.oauth2.server.resource.web.BearerTokenAuthenticationFilter;
-import uk.gov.hmcts.reform.authorisation.filters.ServiceAuthFilter;
-import uk.gov.hmcts.reform.translate.security.JwtGrantedAuthoritiesConverter;
 
-import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
+import java.util.Collection;
+import java.util.List;
+import java.util.function.Function;
+import javax.servlet.http.HttpServletRequest;
 
-@Configuration
 @EnableWebSecurity
-public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+public class SecurityConfiguration {
 
-    @Value("${spring.security.oauth2.client.provider.oidc.issuer-uri}")
-    private String issuerUri;
+  @Value("#{'${idam.s2s-authorised.services}'.split(',')}")
+  private List<String> s2sNamesWhiteList;
 
-    @Value("${oidc.issuer}")
-    private String issuerOverride;
+  @Value("${spring.security.oauth2.client.provider.oidc.issuer-uri}")
+  private String issuerUri;
 
-    private final ServiceAuthFilter serviceAuthFilter;
-    private final JwtAuthenticationConverter jwtAuthenticationConverter;
+  @Value("${oidc.issuer}")
+  private String issuerOverride;
 
-    private static final String[] AUTH_ALLOWED_LIST = {
-        "/swagger-resources/**",
-        "/swagger-ui/**",
-        "/webjars/**",
-        "/v3/api-docs/**",
-        "/health",
-        "/health/liveness",
-        "/health/readiness",
-        "/info",
-        "/favicon.ico",
-        "/"
-    };
+  @Bean
+  public Function<HttpServletRequest, Collection<String>> authorizedServicesExtractor() {
+    return any -> s2sNamesWhiteList;
+  }
 
-    @Autowired
-    public SecurityConfiguration(final ServiceAuthFilter serviceAuthFilter,
-                                 final JwtGrantedAuthoritiesConverter jwtGrantedAuthoritiesConverter) {
-        super();
-        this.serviceAuthFilter = serviceAuthFilter;
-        jwtAuthenticationConverter = new JwtAuthenticationConverter();
-        jwtAuthenticationConverter.setJwtGrantedAuthoritiesConverter(jwtGrantedAuthoritiesConverter);
-    }
+  @Bean
+  JwtDecoder jwtDecoder() {
+    NimbusJwtDecoder jwtDecoder = JwtDecoders.fromOidcIssuerLocation(issuerUri);
+    // We are using issuerOverride instead of issuerUri as SIDAM has the wrong issuer at the moment
+    final OAuth2TokenValidator<Jwt> withTimestamp = new JwtTimestampValidator();
+    final OAuth2TokenValidator<Jwt> validator = new DelegatingOAuth2TokenValidator<>(withTimestamp);
+    jwtDecoder.setJwtValidator(validator);
 
-    @Override
-    public void configure(WebSecurity web) {
-        web.ignoring().antMatchers(AUTH_ALLOWED_LIST);
-    }
+    return jwtDecoder;
+  }
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
-            .addFilterBefore(serviceAuthFilter, BearerTokenAuthenticationFilter.class)
-            .sessionManagement().sessionCreationPolicy(STATELESS).and()
-            .csrf().disable()
-            .formLogin().disable()
-            .logout().disable()
-            .authorizeRequests()
-            .anyRequest()
-            .authenticated()
-            .and()
-            .oauth2ResourceServer()
-            .jwt()
-            .jwtAuthenticationConverter(jwtAuthenticationConverter)
-            .and()
-            .and()
-            .oauth2Client();
-    }
-
-    @Bean
-    JwtDecoder jwtDecoder() {
-        NimbusJwtDecoder jwtDecoder = (NimbusJwtDecoder) JwtDecoders.fromOidcIssuerLocation(issuerUri);
-        OAuth2TokenValidator<Jwt> withTimestamp = new JwtTimestampValidator();
-        OAuth2TokenValidator<Jwt> withIssuer = new JwtIssuerValidator(issuerOverride);
-        OAuth2TokenValidator<Jwt> validator = new DelegatingOAuth2TokenValidator<>(withTimestamp);
-        jwtDecoder.setJwtValidator(validator);
-        return jwtDecoder;
-    }
 }
