+++ src/main/java/uk/gov/hmcts/reform/translate/config/WhiteList.java
@@ -1,0 +1,18 @@
+package uk.gov.hmcts.reform.translate.config;
+
+public final class WhiteList {
+
+    static final String[] AUTH_ALLOWED_LIST = {
+        "/swagger-resources/**",
+        "/swagger-ui/**",
+        "/webjars/**",
+        "/v3/api-docs/**",
+        "/health",
+        "/health/liveness",
+        "/health/readiness",
+        "/info",
+        "/favicon.ico",
+        "/"
+    };
+
+}
